[Decomp][Player] Extract spell-cooldown system into SpellCooldownMgr
