+++ src/game/Object/Player.cpp
@@ -142,7 +142,6 @@ enum CharacterCustomizeFlags
 };
 
 
-
 //== PlayerTaxi ================================================
 
 PlayerTaxi::PlayerTaxi()
@@ -526,7 +525,7 @@ UpdateMask Player::updateVisualBits;
  *
  * @param session The owning world session.
  */
-Player::Player(WorldSession* session): Unit(), m_honorMgr(this), m_glyphMgr(this), m_runeMgr(this), m_mover(this), m_camera(this), m_petMgr(this), m_achievementMgr(this), m_reputationMgr(this)
+Player::Player(WorldSession* session): Unit(), m_honorMgr(this), m_spellCooldownMgr(this), m_glyphMgr(this), m_runeMgr(this), m_mover(this), m_camera(this), m_petMgr(this), m_achievementMgr(this), m_reputationMgr(this)
 {
     m_transport = 0;
 
@@ -2238,9 +2237,6 @@ void Player::RemoveFromWorld()
 }
 
 
-
-
-
 /**
  * @brief Gets an NPC the player can currently interact with.
  *
@@ -3006,7 +3002,7 @@ void Player::SendInitialSpells()
      * * * * * * * * * * * * * * * * */
     uint16 spellCount = 0;
 
-    WorldPacket data(SMSG_INITIAL_SPELLS, (1 + 2 + 4 * m_spells.size() + 2 + m_spellCooldowns.size() * (2 + 2 + 2 + 4 + 4)));
+    WorldPacket data(SMSG_INITIAL_SPELLS, (1 + 2 + 4 * m_spells.size() + 2 + GetSpellCooldownMap().size() * (2 + 2 + 2 + 4 + 4)));
     data << uint8(0);
 
     /* * * * * * * * * * * * * * * * *
@@ -3042,9 +3038,9 @@ void Player::SendInitialSpells()
     data.put<uint16>(countPos, spellCount);                 // write real count value
 
     /* For each spell the player has on cooldown */
-    uint16 spellCooldowns = m_spellCooldowns.size();
+    uint16 spellCooldowns = GetSpellCooldownMap().size();
     data << uint16(spellCooldowns);
-    for (SpellCooldowns::const_iterator itr = m_spellCooldowns.begin(); itr != m_spellCooldowns.end(); ++itr)
+    for (SpellCooldowns::const_iterator itr = GetSpellCooldownMap().begin(); itr != GetSpellCooldownMap().end(); ++itr)
     {
         /* If the spell doesn't exist in the spellbook, just ignore it */
         SpellEntry const* sEntry = sSpellStore.LookupEntry(itr->first);
@@ -3086,11 +3082,6 @@ void Player::SendInitialSpells()
 
     DETAIL_LOG("CHARACTER: Sent Initial Spells");
 }
-
-
-
-
-
 
 
 /**
@@ -3392,12 +3383,6 @@ void Player::DeleteOldCharacters(uint32 keepDays)
 }
 
 
-
-
-
-
-
-
 /**
  * @brief Attempts to improve defense skill and refresh defense-derived bonuses.
  */
@@ -3674,9 +3659,6 @@ void Player::setFactionForRace(uint8 race)
 }
 
 
-
-
-
 void Player::UpdateArenaFields(void)
 {
     /* arena calcs go here */
@@ -3727,7 +3709,6 @@ void Player::ModifyArenaPoints(int32 value)
 }
 
 
-
 /**
  * @brief Checks whether the player is eligible to interact with a capture point.
  *
@@ -3745,15 +3726,7 @@ bool Player::CanUseCapturePoint()
 }
 
 
-
 //---------------------------------------------------------//
-
-
-
-
-
-
-
 
 
 /**
@@ -3768,20 +3741,6 @@ bool Player::CanUseCapturePoint()
  * functions for item final remove/destroy from inventory. If new RemoveItem calls added need be sure that
  * function will call after it in some way if need.
  */
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 
 
 /**
@@ -3899,8 +3858,6 @@ void Player::SendInitWorldStates(uint32 zoneid, uint32 areaid)
 }
 
 
-
-
 /**
  * @brief Sends a bind point confirmation prompt to the client.
  *
@@ -3952,17 +3909,6 @@ void Player::SendPetSkillWipeConfirm()
  * @param i The virtual slot index.
  * @param item The item to reflect in the virtual slot.
  */
-
-
-
-
-
-
-
-
-
-
-
 
 
 /*********************************************************/
@@ -4103,206 +4049,6 @@ void Player::InitDisplayIds()
 }
 
 
-
-
-
-
-/**
- * @brief Applies personal and category cooldowns for a spell cast.
- *
- * @param spellInfo The spell entry that triggered the cooldown.
- * @param itemId The casting item entry, if any.
- * @param spell The active spell instance.
- * @param infinityCooldown True to apply a long-lived cooldown marker.
- */
-void Player::AddSpellAndCategoryCooldowns(SpellEntry const* spellInfo, uint32 itemId, Spell* spell, bool infinityCooldown)
-{
-    // init cooldown values
-    uint32 cat   = 0;
-    int32 rec    = -1;
-    int32 catrec = -1;
-
-    // some special item spells without correct cooldown in SpellInfo
-    // cooldown information stored in item prototype
-    // This used in same way in WorldSession::HandleItemQuerySingleOpcode data sending to client.
-
-    if (itemId)
-    {
-        if (ItemPrototype const* proto = ObjectMgr::GetItemPrototype(itemId))
-        {
-            for (int idx = 0; idx < MAX_ITEM_PROTO_SPELLS; ++idx)
-            {
-                if (proto->Spells[idx].SpellId == spellInfo->Id)
-                {
-                    cat    = proto->Spells[idx].SpellCategory;
-                    rec    = proto->Spells[idx].SpellCooldown;
-                    catrec = proto->Spells[idx].SpellCategoryCooldown;
-                    break;
-                }
-            }
-        }
-    }
-
-    // if no cooldown found above then base at DBC data
-    if (rec < 0 && catrec < 0)
-    {
-        cat = spellInfo->Category;
-        rec = spellInfo->RecoveryTime;
-        catrec = spellInfo->CategoryRecoveryTime;
-    }
-
-    time_t curTime = time(NULL);
-
-    time_t catrecTime;
-    time_t recTime;
-
-    // overwrite time for selected category
-    if (infinityCooldown)
-    {
-        // use +MONTH as infinity mark for spell cooldown (will checked as MONTH/2 at save ans skipped)
-        // but not allow ignore until reset or re-login
-        catrecTime = catrec > 0 ? curTime + infinityCooldownDelay : 0;
-        recTime    = rec    > 0 ? curTime + infinityCooldownDelay : catrecTime;
-    }
-    else
-    {
-        // shoot spells used equipped item cooldown values already assigned in GetAttackTime(RANGED_ATTACK)
-        // prevent 0 cooldowns set by another way
-        if (rec <= 0 && catrec <= 0 && (cat == 76 || (IsAutoRepeatRangedSpell(spellInfo) && spellInfo->Id != SPELL_ID_AUTOSHOT)))
-        {
-            rec = GetAttackTime(RANGED_ATTACK);
-        }
-
-        // Now we have cooldown data (if found any), time to apply mods
-        if (rec > 0)
-        {
-            ApplySpellMod(spellInfo->Id, SPELLMOD_COOLDOWN, rec);
-        }
-
-        if (catrec > 0)
-        {
-            ApplySpellMod(spellInfo->Id, SPELLMOD_COOLDOWN, catrec);
-        }
-
-        // replace negative cooldowns by 0
-        if (rec < 0) rec = 0;
-        {
-            if (catrec < 0) catrec = 0;
-        }
-
-        // no cooldown after applying spell mods
-        if (rec == 0 && catrec == 0)
-        {
-            return;
-        }
-
-        catrecTime = catrec ? curTime + catrec / IN_MILLISECONDS : 0;
-        recTime    = rec ? curTime + rec / IN_MILLISECONDS : catrecTime;
-    }
-
-    // self spell cooldown
-    if (recTime > 0)
-    {
-        AddSpellCooldown(spellInfo->Id, itemId, recTime);
-    }
-
-    // category spells
-    if (cat && catrec > 0)
-    {
-        SpellCategoryStore::const_iterator i_scstore = sSpellCategoryStore.find(cat);
-        if (i_scstore != sSpellCategoryStore.end())
-        {
-            for (SpellCategorySet::const_iterator i_scset = i_scstore->second.begin(); i_scset != i_scstore->second.end(); ++i_scset)
-            {
-                if (*i_scset == spellInfo->Id)              // skip main spell, already handled above
-                {
-                    continue;
-                }
-
-                AddSpellCooldown(*i_scset, itemId, catrecTime);
-            }
-        }
-    }
-}
-
-/**
- * @brief Stores a cooldown entry for a spell.
- *
- * @param spellid The spell identifier.
- * @param itemid The associated item identifier, if any.
- * @param end_time The server time when the cooldown ends.
- */
-void Player::AddSpellCooldown(uint32 spellid, uint32 itemid, time_t end_time)
-{
-    SpellCooldown sc;
-    sc.end = end_time;
-    sc.itemid = itemid;
-    m_spellCooldowns[spellid] = sc;
-}
-
-/**
- * @brief Applies cooldowns and notifies the client about a spell cooldown event.
- *
- * @param spellInfo The spell entry that triggered the cooldown.
- * @param itemId The associated item entry, if any.
- * @param spell The active spell instance.
- */
-void Player::SendCooldownEvent(SpellEntry const* spellInfo, uint32 itemId, Spell* spell)
-{
-    // start cooldowns at server side, if any
-    AddSpellAndCategoryCooldowns(spellInfo, itemId, spell);
-
-    // Send activate cooldown timer (possible 0) at client side
-    WorldPacket data(SMSG_COOLDOWN_EVENT, (4 + 8));
-    data << uint32(spellInfo->Id);
-    data << GetObjectGuid();
-    SendDirectMessage(&data);
-}
-
-void Player::UpdatePotionCooldown(Spell* spell)
-{
-    // no potion used in combat or still in combat
-    if (!m_lastPotionId || IsInCombat())
-    {
-        return;
-    }
-
-    // Call not from spell cast, send cooldown event for item spells if no in combat
-    if (!spell)
-    {
-        // spell/item pair let set proper cooldown (except nonexistent charged spell cooldown spellmods for potions)
-        if (ItemPrototype const* proto = ObjectMgr::GetItemPrototype(m_lastPotionId))
-            for (int idx = 0; idx < 5; ++idx)
-            {
-                if (proto->Spells[idx].SpellId && proto->Spells[idx].SpellTrigger == ITEM_SPELLTRIGGER_ON_USE)
-                    if (SpellEntry const* spellInfo = sSpellStore.LookupEntry(proto->Spells[idx].SpellId))
-                    {
-                        SendCooldownEvent(spellInfo, m_lastPotionId);
-                    }
-            }
-    }
-    // from spell cases (m_lastPotionId set in Spell::SendSpellCooldown)
-    else
-    {
-        SendCooldownEvent(spell->m_spellInfo, m_lastPotionId, spell);
-    }
-
-    m_lastPotionId = 0;
-}
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 void Player::SetPhaseMask(uint32 newPhaseMask, bool update)
 {
     // GM-mode have mask PHASEMASK_ANYWHERE always
@@ -4328,7 +4074,6 @@ void Player::InitPrimaryProfessions()
 {
     SetFreePrimaryProfessions(sWorld.getConfig(CONFIG_UINT32_MAX_PRIMARY_TRADE_SKILL));
 }
-
 
 
 void Player::SendInitialPacketsBeforeAddToMap()
@@ -4436,7 +4181,6 @@ void Player::SendInitialPacketsAfterAddToMap()
 }
 
 
-
 /**
  * @brief Applies the default equip cooldown for item use spells.
  *
@@ -4475,10 +4219,6 @@ void Player::ApplyEquipCooldown(Item* pItem)
 }
 
 
-
-
-
-
 /**
  * @brief Sends visible aura duration updates for a target to the player.
  *
@@ -4506,8 +4246,6 @@ void Player::SendAurasForTarget(Unit* target)
 
     GetSession()->SendPacket(&data);
 }
-
-
 
 
 /**
@@ -4898,10 +4636,6 @@ uint32 Player::GetResurrectionSpellId()
 }
 
 
-
-
-
-
 /**
  * @brief Gets the player's base weapon skill for an attack type.
  *
@@ -4982,10 +4716,6 @@ void Player::SetClientControl(Unit* target, uint8 allowMove)
     data << uint8(allowMove);
     GetSession()->SendPacket(&data);
 }
-
-
-
-
 
 
 /**
@@ -5812,7 +5542,6 @@ SpellEntry const* Player::GetKnownTalentRankById(int32 talentId) const
 }
 
 
-
 void Player::UpdateKnownCurrencies(uint32 itemId, bool apply)
 {
     if (CurrencyTypesEntry const* ctEntry = sCurrencyTypesStore.LookupEntry(itemId))
@@ -5940,8 +5669,6 @@ void Player::_SaveBGData()
 
     m_bgData.m_needSave = false;
 }
-
-
 
 
 /**
@@ -6159,10 +5886,6 @@ Object* Player::GetObjectByTypeMask(ObjectGuid guid, TypeMask typemask)
 }
 
 
-
-
-
-
 float Player::GetCollisionHeight(bool mounted) const
 {
     if (mounted)
@@ -6238,14 +5961,4 @@ void Player::_LoadRandomBGStatus(QueryResult *result)
         delete result;
     }
 }
-
-
-
-
-
-
-
-
-
-
